Set exit code on failed tests despite `should_exit` option
When running gut from command line, it becomes possible to accidentaly
push changes that didn't pass the tests when using continuous integration
if `should_exit` option is not set.

This fixes the described issue by ensuring that the exit code is set regardless
of possible options. It's also not necessary to set exit code to zero because
the program exits with zero by default.