Add maximize option to command line and the editor
This allows to fit test runner window to the viewport. The maximized mode
can be toggled from the editor, from a command line and be fetched from
the gut configuration file.

Also defined `title_offset` variable to automatically fetch the window's
title bar height that was hardcoded and used for positioning the window.