Automatically enable exclude_addons option when running tests
GUT will throw errors if this option is disabled, e.g. due to overriding
native methods.
When working on addons it is usually desired to disable exclude_addons
in order to get proper code diagnostics.
This however makes it tedious to run GUT because the option has to be
manually disabled before running tests and manually enabled again
afterwards to get diagnostics.
To workaround this issue, this commit automatically enables the
exclude_addons option when running tests, so regardless of the project
settings in the editor, GUT will always run as intended.